requirements.txt
boto==2.42.0
dj-database-url==0.4.1
Django==1.10
django-crispy-forms==1.6.0
django-storages==1.5.0
djangorestframework==3.4.1
gunicorn==19.6.0
Pillow==3.3.0
psycopg2==2.6.2
whitenoise==3.2
